pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-summarizer"
version = "0.1.0"
description = "A tool for analyzing YouTube video transcripts using LLM models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "litellm",      
    "youtube-transcript-api",
    "langchain-community",
    "langchain-openai",
    "langchain-core",
    "langchain-litellm",
    "langchain-google-genai",
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/AlexanderSlivinskiy/youtubeSummarizer"
Repository = "https://github.com/AlexanderSlivinskiy/youtubeSummarizer"

[project.scripts]
youtube-summarizer = "youtube_summarizer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["youtube_summarizer*"]
